% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomcolour.R
\docType{package}
\name{randomcolour-package}
\alias{randomcolour}
\alias{randomcolour-package}
\title{randomcolour: Random colour generation and presentation}
\description{
Random colour generation in R, for use in anywhere colours are needed.
}
\details{
randomcolour: A package for exploring the techniques and best practices
in R through the minimal example of random colour generation.

To learn more about randomcolour, start with the vignettes:
`browseVignettes(package = "randomcolour")`
}
\author{
\strong{Maintainer}: Matt Malin \email{matt@randomcolour.com}

Authors:
\itemize{
  \item Jonathan Ng \email{jonathan@datastrategywithjonathan.com}
}

}
\keyword{internal}
