% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomcolour.R
\name{randomcolour}
\alias{randomcolour}
\alias{randomcolor}
\title{Generate a random colour}
\usage{
randomcolour()

randomcolor()
}
\value{
A hexidecimal string of form #XXXXXX representing a random colour across the 24 bit colour range.

Mimics the randomcolour.com approach for selecting a random integer between 0 and 1677715 and expressing
as a hexidecimal string of appropriate format.
}
\description{
Generate a random colour
}
\section{Functions}{
\itemize{
\item \code{randomcolour}: main function

\item \code{randomcolor}: alternative spelling, calls randomcolour()
}}

\examples{
randomcolour()

randomcolor()

}
